Move Factory class to factory module
