+++ plextraktsync/factory/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from plextraktsync.util.Factory import Factory
+from .Factory import Factory
 
 factory = Factory()
 logger = factory.logger
